Moved related code of each macro into individual corresponding files
